Improve Payout Calculations

+++ Domain/Players/Player.cs
@@ -3,5 +3,5 @@ namespace Domain.Players;
 public record Player
 {
     public required string Name { get; init; }
-    public required decimal Balance { get; set; }
+    public required int Balance { get; set; }
 }
